+++ R/workbook_column_widths.R
@@ -136,23 +136,33 @@ Workbook$methods(setColWidths = function(sheet) {
     widths[missingAuto] <- 9.15
   }
 
-  # Check if any conflicting column outline levels
+  # Check if any conflicting existing levels
   if (any(cols %in% names(worksheets[[sheet]]$cols))) {
+
     for (i in intersect(cols, names(worksheets[[sheet]]$cols))) {
-      width_hidden <- attr(wb$colWidths[[sheet]], "hidden")[attr(wb$colWidths[[sheet]], "names") == i]
-      width_widths <- attr(wb$colWidths[[sheet]], "widths")[attr(wb$colWidths[[sheet]], "names") == i]
-      worksheets[[sheet]]$cols[[i]] <<- sub("((?<=hidden=\")(\\w)\")", paste0(width_hidden, "\" width=\"", width_widths, "\" customWidth=\"1\"/>"), worksheets[[sheet]]$cols[[i]], perl = TRUE)
+      
+      width_hidden <- attr(colWidths[[sheet]], "hidden")[attr(colWidths[[sheet]], "names") == i]
+      width_widths <- as.numeric(colWidths[[sheet]][attr(colWidths[[sheet]], "names") == i]) + 0.71
+
+      # If column already has a custom width, just update the width and hidden attributes
+      if (grepl("customWidth", worksheets[[sheet]]$cols[[i]])) {
+        worksheets[[sheet]]$cols[[i]] <<- sub('(width=\\").*?(\\"\\shidden=\\").*?(\\")', paste0("\\1", width_widths, "\\2", width_hidden, "\\3"), worksheets[[sheet]]$cols[[i]], perl = TRUE)
+      } else {
+      # If column exists, but doesn't have a custom width
+        worksheets[[sheet]]$cols[[i]] <<- sub("((?<=hidden=\")(\\w)\")", paste0(width_hidden, "\" width=\"", width_widths, "\" customWidth=\"1\"/>"), worksheets[[sheet]]$cols[[i]], perl = TRUE)
+      }
     }
 
     cols <- cols[!cols %in% names(worksheets[[sheet]]$cols)]
   }
 
-  ## Calculate width of auto
-  colNodes <- sprintf('<col min="%s" max="%s" width="%s" hidden="%s" customWidth="1"/>', cols, cols, widths, hidden)
-
-  ## Append new col widths XML to worksheets[[sheet]]$cols
-  worksheets[[sheet]]$cols <<- append(worksheets[[sheet]]$cols, colNodes)
-  names(worksheets[[sheet]]$cols) <<- cols
+  # Add remaining columns
+  if (length(cols) > 0) {
+    colNodes <- sprintf('<col min="%s" max="%s" width="%s" hidden="%s" customWidth="1"/>', cols, cols, widths, hidden)
+    names(colNodes) <- cols
+    worksheets[[sheet]]$cols <<- append(worksheets[[sheet]]$cols, colNodes)
+  }
+  
 })